Support `card-mod` by meeting its expectations
`card-mod` expects that the `<ha-card/>`, or one of its ancestors, has
an attribute named `config` or `_config` where the initial card
configuration is stored. It searches for a `card_mod:` section in it.

`plotly-graph` uses such a `config` attribute, but stores its own
internal understanding of what the YAML config meant, which means in
particular that the `card_mod:` section is filtered out (and for that
matter any configuration that `plotly-graph` is not aware of).

Store the vetted configuration in `parsed_config` instead, and remember
the initial unfiltered configuration in the `config` attribute for
`card_mod` and others to use.

Fixes #86.